4 tests, 16 assertions in test_get.rb, issue #14

+++ test/test_get.rb
@@ -4,23 +4,22 @@ require 'rouster'
 require 'rouster/tests'
 require 'test/unit'
 
-@app = Rouster.new(:name => 'app')
-@ppm = Rouster.new(:name => 'ppm', :sudo => false)
-
-@app.up()
-
 class TestPut < Test::Unit::TestCase
 
   def setup
-    #@app = Rouster.new({:name => 'app'})
-    #@ppm = Rouster.new({:name => 'ppm', :sudo => false})
-    #@app.up()
-    #@ppm.up()
-    @kg_local_location  = sprintf('/tmp/rouster-test_get_local.%s.%s', $$, Time.now.to_i)
-    @kg_remote_location = '/etc/hosts'
-    @kb_dne_location = '/tmp/this-doesnt_exist/and/never/will.txt'
+    @app = Rouster.new(:name => 'app')
 
-    File.delete(@kg_local_location)
+    @app.up()
+
+    @kg_local_location  = sprintf('/tmp/rouster-test_get_local.%s.%s', $$, Time.now.to_i)
+    @kg_local_location.freeze
+    @kg_remote_location = '/etc/hosts'
+    @kb_dne_location    = '/tmp/this-doesnt_exist/and/never/will.txt'
+
+    File.delete(@kg_local_location) if File.file?(@kg_local_location).true?
+
+    assert_equal(@app.is_available_via_ssh?, true, 'app is available via SSH')
+    assert_equal(File.file?(@kg_local_location), false, 'test KG file not present')
   end
 
   def test_happy_path
@@ -29,33 +28,45 @@ class TestPut < Test::Unit::TestCase
       @app.get(@kg_remote_location, @kg_local_location)
     end
 
-    assert(@app.is_file?(@kg_local_location))
+    assert(File.file?(@kg_local_location))
   end
 
   def test_local_path_dne
 
-    assert_raise FileTransferError do
+    assert_raise Rouster::FileTransferError do
       @app.get(@kg_remote_location, @kb_dne_location)
     end
 
-    assert_equal(false, @app.is_file?(@kg_location), 'known bad local path DNE')
+    assert_equal(false, File.file?(@kg_local_location), 'known bad local path DNE')
   end
 
   def test_remote_path_dne
 
-    assert_raise SSHConnectionError do
+    assert_raise Rouster::FileTransferError do
       @app.get(@kb_dne_location, @kg_local_location)
     end
 
-    assert_equal(false, @app.is_file?(@kg_local_location), 'known bad remote file path DNE')
+    assert_equal(false, File.file?(@kg_local_location), 'known bad remote file path DNE')
 
   end
+
+  def test_with_suspended_machine
+    @app.suspend()
+
+    assert_raise Rouster::SSHConnectionError do
+      @app.get(@kg_remote_location, @kg_local_location)
+    end
+
+    assert_equal(false, File.file?(@kg_local_location), 'when machine is suspended, unable to get from it')
+  end
+
+
 
   def teardown
     # TODO we should suspend instead if any test failed for triage
     #@app.destroy()
     #@ppm.destroy()
 
-    File.delete(@kg_local_location)
+    File.delete(@kg_local_location) if File.file?(@kg_local_location).true?
   end
 end
